First version of more stand-alone iframe content servlet
Update to match merged changes

Revert unnecessary changes, strip down admin functionality from original iframe-servlet

Make permission overriding more obvious, comments
