+++ spark-confluence/src/main/java/com/k15t/spark/confluence/ConfluenceIframeAdminAppAction.java
@@ -36,12 +36,6 @@ public abstract class ConfluenceIframeAdminAppAction extends ConfluenceActionSup
 
 
     @Override
-    public String getIframeContextInitializedCallbackName() {
-        return "contextInitializedCallback";
-    }
-
-
-    @Override
     public String getSpaQueryString() {
         return ServletActionContext.getRequest().getQueryString();
     }
